third_party/WebKit/LayoutTests/plugins/fullscreen-plugins-dont-reload-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
go fullscreen
There should only be one ALERT. If there were two, the plugin was reloaded during the transition to fullscreen.
